Show correct answer in question, question authentication changes

+++ src/main/java/sk/rzapach/advent/web/rest/QuestionResource.java
@@ -1,23 +1,26 @@
 package sk.rzapach.advent.web.rest;
 
-import sk.rzapach.advent.domain.Question;
-import sk.rzapach.advent.service.QuestionService;
-import sk.rzapach.advent.web.rest.errors.BadRequestAlertException;
-import sk.rzapach.advent.service.dto.QuestionCriteria;
-import sk.rzapach.advent.service.QuestionQueryService;
-
 import io.github.jhipster.web.util.HeaderUtil;
 import io.github.jhipster.web.util.ResponseUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import sk.rzapach.advent.domain.Question;
+import sk.rzapach.advent.domain.User;
+import sk.rzapach.advent.security.AuthoritiesConstants;
+import sk.rzapach.advent.service.QuestionQueryService;
+import sk.rzapach.advent.service.QuestionService;
+import sk.rzapach.advent.service.UserService;
+import sk.rzapach.advent.service.dto.QuestionCriteria;
+import sk.rzapach.advent.web.rest.errors.BadRequestAlertException;
 
 import javax.validation.Valid;
 import java.net.URI;
 import java.net.URISyntaxException;
-
+import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 
@@ -36,12 +39,21 @@ public class QuestionResource {
     private String applicationName;
 
     private final QuestionService questionService;
-
     private final QuestionQueryService questionQueryService;
+    private final UserService userService;
 
-    public QuestionResource(QuestionService questionService, QuestionQueryService questionQueryService) {
+    public QuestionResource(QuestionService questionService, QuestionQueryService questionQueryService, UserService userService) {
         this.questionService = questionService;
         this.questionQueryService = questionQueryService;
+        this.userService = userService;
+    }
+
+    private User getLoggedUser() {
+        return userService.getUserWithAuthorities().get();
+    }
+
+    private boolean isAdmin(User user) {
+        return user.getAuthorities().stream().anyMatch(a -> AuthoritiesConstants.ADMIN.equals(a.getName()));
     }
 
     /**
@@ -57,6 +69,11 @@ public class QuestionResource {
         if (question.getId() != null) {
             throw new BadRequestAlertException("A new question cannot already have an ID", ENTITY_NAME, "idexists");
         }
+
+        if (!isAdmin(getLoggedUser())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         Question result = questionService.save(question);
         return ResponseEntity.created(new URI("/api/questions/" + result.getId()))
             .headers(HeaderUtil.createEntityCreationAlert(applicationName, false, ENTITY_NAME, result.getId().toString()))
@@ -75,6 +92,11 @@ public class QuestionResource {
     @PutMapping("/questions")
     public ResponseEntity<Question> updateQuestion(@Valid @RequestBody Question question) throws URISyntaxException {
         log.debug("REST request to update Question : {}", question);
+
+        if (!isAdmin(getLoggedUser())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         if (question.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
@@ -87,7 +109,6 @@ public class QuestionResource {
     /**
      * {@code GET  /questions} : get all the questions.
      *
-
      * @param criteria the criteria which the requested entities should match.
      * @return the {@link ResponseEntity} with status {@code 200 (OK)} and the list of questions in body.
      */
@@ -95,18 +116,24 @@ public class QuestionResource {
     public ResponseEntity<List<Question>> getAllQuestions(QuestionCriteria criteria) {
         log.debug("REST request to get Questions by criteria: {}", criteria);
         List<Question> entityList = questionQueryService.findByCriteria(criteria);
+        entityList.forEach(this::sanitizeQuestion);
         return ResponseEntity.ok().body(entityList);
     }
 
     /**
-    * {@code GET  /questions/count} : count all the questions.
-    *
-    * @param criteria the criteria which the requested entities should match.
-    * @return the {@link ResponseEntity} with status {@code 200 (OK)} and the count in body.
-    */
+     * {@code GET  /questions/count} : count all the questions.
+     *
+     * @param criteria the criteria which the requested entities should match.
+     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and the count in body.
+     */
     @GetMapping("/questions/count")
     public ResponseEntity<Long> countQuestions(QuestionCriteria criteria) {
         log.debug("REST request to count Questions by criteria: {}", criteria);
+
+        if (!isAdmin(getLoggedUser())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         return ResponseEntity.ok().body(questionQueryService.countByCriteria(criteria));
     }
 
@@ -120,6 +147,7 @@ public class QuestionResource {
     public ResponseEntity<Question> getQuestion(@PathVariable Long id) {
         log.debug("REST request to get Question : {}", id);
         Optional<Question> question = questionService.findOne(id);
+        question.ifPresent(this::sanitizeQuestion);
         return ResponseUtil.wrapOrNotFound(question);
     }
 
@@ -132,7 +160,24 @@ public class QuestionResource {
     @DeleteMapping("/questions/{id}")
     public ResponseEntity<Void> deleteQuestion(@PathVariable Long id) {
         log.debug("REST request to delete Question : {}", id);
+
+        if (!isAdmin(getLoggedUser())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         questionService.delete(id);
         return ResponseEntity.noContent().headers(HeaderUtil.createEntityDeletionAlert(applicationName, false, ENTITY_NAME, id.toString())).build();
+    }
+
+    private void sanitizeQuestion(Question question) {
+        if (!isAdmin(getLoggedUser())) {
+            if (Instant.now().isBefore(question.getTime())) {
+                question.setText("");
+                question.setChoices("");
+            }
+            if (!question.isShowAnswer()) {
+                question.setAnswer("");
+            }
+        }
     }
 }
